@@ -11,7 +11,7 @@ setup(
     author="Tskit Developers",
     author_email="jerome.kelleher@well.ox.ac.uk",
     url="http://pypi.python.org/pypi/tszip",
-    setup_requires=['setuptools_scm'],
+    setup_requires=["setuptools_scm"],
     install_requires=[
         "numpy",
         "humanize",
@@ -22,14 +22,14 @@ setup(
     keywords=["Tree sequences", "tskit"],
     license="MIT",
     platforms=["POSIX", "Windows", "MacOS X"],
-    python_requires=">=3.4",
+    python_requires=">=3.7",
     classifiers=[
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
-        "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
-        'Programming Language :: Python :: 3 :: Only',
-        'License :: OSI Approved :: MIT License',
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3 :: Only",
+        "License :: OSI Approved :: MIT License",
         "Development Status :: 3 - Alpha",
         "Environment :: Other Environment",
         "Intended Audience :: Science/Research",
@@ -38,13 +38,13 @@ setup(
         "Topic :: Scientific/Engineering :: Bio-Informatics",
     ],
     project_urls={
-        'Bug Reports': 'https://github.com/tskit-dev/tszip/issues',
-        'Source': 'https://github.com/tskit-dev/tszip',
+        "Bug Reports": "https://github.com/tskit-dev/tszip/issues",
+        "Source": "https://github.com/tskit-dev/tszip",
     },
     entry_points={
-        'console_scripts': [
-            'tszip=tszip.cli:tszip_main',
-            'tsunzip=tszip.cli:tsunzip_main',
+        "console_scripts": [
+            "tszip=tszip.cli:tszip_main",
+            "tsunzip=tszip.cli:tsunzip_main",
         ]
     },
     use_scm_version={"write_to": "tszip/_version.py"},
