@@ -22,20 +22,19 @@
 """
 Compression utilities for tskit tree sequences.
 """
-import logging
-import os
-import warnings
 import contextlib
-import zipfile
-import tempfile
-import os.path
 import json
+import logging
+import os.path
+import tempfile
+import warnings
+import zipfile
 
+import humanize
 import numcodecs
-import zarr
 import numpy as np
 import tskit
-import humanize
+import zarr
 
 from . import exceptions
 from . import provenance
@@ -55,13 +54,13 @@ def minimal_dtype(array):
     dtype = array.dtype
     if array.shape[0] == 0:
         return dtype
-    if dtype.kind == 'u':
+    if dtype.kind == "u":
         maxval = np.max(array)
         dtypes = [np.uint8, np.uint16, np.uint32, np.uint64]
         for dtype in map(np.dtype, dtypes):
             if maxval <= np.iinfo(dtype).max:
                 break
-    elif dtype.kind == 'i':
+    elif dtype.kind == "i":
         minval = np.min(array)
         maxval = np.max(array)
         dtypes = [np.int8, np.int16, np.int32, np.int64]
@@ -91,12 +90,12 @@ def compress(ts, path, variants_only=False):
     destdir = os.path.dirname(os.path.abspath(destination))
     with tempfile.TemporaryDirectory(dir=destdir, prefix=".tszip_work_") as tmpdir:
         filename = os.path.join(tmpdir, "tmp.trees.tgz")
-        logging.debug("Writing to temporary file {}".format(filename))
-        with zarr.ZipStore(filename, mode='w') as store:
+        logging.debug(f"Writing to temporary file {filename}")
+        with zarr.ZipStore(filename, mode="w") as store:
             root = zarr.group(store=store)
             compress_zarr(ts, root, variants_only=variants_only)
         os.replace(filename, destination)
-    logging.info("Wrote {}".format(destination))
+    logging.info(f"Wrote {destination}")
 
 
 def decompress(path):
@@ -113,10 +112,11 @@ def decompress(path):
         return decompress_zarr(root)
 
 
-class Column(object):
+class Column:
     """
     A single column that is stored in the compressed output.
     """
+
     def __init__(self, name, array, delta_filter=False):
         self.name = name
         self.array = array
@@ -126,22 +126,30 @@ class Column(object):
         shape = self.array.shape
         chunks = shape
         if shape[0] == 0:
-            chunks = 1,
+            chunks = (1,)
         dtype = minimal_dtype(self.array)
         filters = None
         if self.delta_filter:
             filters = [numcodecs.Delta(dtype=dtype)]
         compressed_array = root.empty(
-            self.name, chunks=chunks, shape=shape, dtype=dtype,
-            filters=filters, compressor=compressor)
+            self.name,
+            chunks=chunks,
+            shape=shape,
+            dtype=dtype,
+            filters=filters,
+            compressor=compressor,
+        )
         compressed_array[:] = self.array
         ratio = 0
         if compressed_array.nbytes > 0:
             ratio = compressed_array.nbytes / compressed_array.nbytes_stored
-        logger.debug("{}: output={} compression={:.1f}".format(
-            self.name,
-            humanize.naturalsize(compressed_array.nbytes_stored, binary=True),
-            ratio))
+        logger.debug(
+            "{}: output={} compression={:.1f}".format(
+                self.name,
+                humanize.naturalsize(compressed_array.nbytes_stored, binary=True),
+                ratio,
+            )
+        )
 
 
 def compress_zarr(ts, root, variants_only=False):
@@ -160,9 +168,18 @@ def compress_zarr(ts, root, variants_only=False):
         tables = ts.tables
         node_time = tables.nodes.time
 
-    coordinates = np.unique(np.hstack([
-        [0, ts.sequence_length], tables.edges.left, tables.edges.right,
-        tables.sites.position, tables.migrations.left, tables.migrations.right]))
+    coordinates = np.unique(
+        np.hstack(
+            [
+                [0, ts.sequence_length],
+                tables.edges.left,
+                tables.edges.right,
+                tables.sites.position,
+                tables.migrations.left,
+                tables.migrations.right,
+            ]
+        )
+    )
 
     with warnings.catch_warnings():
         warnings.simplefilter("ignore")
@@ -181,43 +198,46 @@ def compress_zarr(ts, root, variants_only=False):
         Column("individuals/flags", tables.individuals.flags),
         Column("individuals/location", tables.individuals.location),
         Column(
-            "individuals/location_offset", tables.individuals.location_offset,
-            delta_filter=True),
+            "individuals/location_offset",
+            tables.individuals.location_offset,
+            delta_filter=True,
+        ),
         Column("individuals/metadata", tables.individuals.metadata),
         Column(
-            "individuals/metadata_offset", tables.individuals.metadata_offset,
-            delta_filter=True),
-
+            "individuals/metadata_offset",
+            tables.individuals.metadata_offset,
+            delta_filter=True,
+        ),
         Column("nodes/time", node_time),
         Column("nodes/flags", tables.nodes.flags),
         Column("nodes/population", tables.nodes.population),
         Column("nodes/individual", tables.nodes.individual),
         Column("nodes/metadata", tables.nodes.metadata),
         Column(
-            "nodes/metadata_offset", tables.nodes.metadata_offset, delta_filter=True),
-
+            "nodes/metadata_offset", tables.nodes.metadata_offset, delta_filter=True
+        ),
         # Delta filtering makes storage slightly worse for everything except parent.
         Column("edges/left", np.searchsorted(coordinates, tables.edges.left)),
         Column("edges/right", np.searchsorted(coordinates, tables.edges.right)),
         Column("edges/parent", tables.edges.parent, delta_filter=True),
         Column("edges/child", tables.edges.child),
-
         Column("migrations/left", np.searchsorted(coordinates, tables.migrations.left)),
         Column(
-            "migrations/right", np.searchsorted(coordinates, tables.migrations.right)),
+            "migrations/right", np.searchsorted(coordinates, tables.migrations.right)
+        ),
         Column("migrations/node", tables.migrations.node),
         Column("migrations/source", tables.migrations.source),
         Column("migrations/dest", tables.migrations.dest),
         Column("migrations/time", tables.migrations.time),
-
         Column(
-            "sites/position", np.searchsorted(coordinates, tables.sites.position),
-            delta_filter=True),
+            "sites/position",
+            np.searchsorted(coordinates, tables.sites.position),
+            delta_filter=True,
+        ),
         Column("sites/ancestral_state", tables.sites.ancestral_state),
         Column("sites/ancestral_state_offset", tables.sites.ancestral_state_offset),
         Column("sites/metadata", tables.sites.metadata),
         Column("sites/metadata_offset", tables.sites.metadata_offset),
-
         Column("mutations/site", tables.mutations.site),
         Column("mutations/node", tables.mutations.node),
         Column("mutations/parent", tables.mutations.parent),
@@ -225,10 +245,8 @@ def compress_zarr(ts, root, variants_only=False):
         Column("mutations/derived_state_offset", tables.mutations.derived_state_offset),
         Column("mutations/metadata", tables.mutations.metadata),
         Column("mutations/metadata_offset", tables.mutations.metadata_offset),
-
         Column("populations/metadata", tables.populations.metadata),
         Column("populations/metadata_offset", tables.populations.metadata_offset),
-
         Column("provenances/timestamp", tables.provenances.timestamp),
         Column("provenances/timestamp_offset", tables.provenances.timestamp_offset),
         Column("provenances/record", tables.provenances.record),
@@ -240,7 +258,9 @@ def compress_zarr(ts, root, variants_only=False):
     # difference between compression level 6 and 9, and it's extremely fast in any case
     # so there's no point in adding complexity. The shuffle filter in particular makes
     # big difference.
-    compressor = numcodecs.Blosc(cname='zstd', clevel=9, shuffle=numcodecs.Blosc.SHUFFLE)
+    compressor = numcodecs.Blosc(
+        cname="zstd", clevel=9, shuffle=numcodecs.Blosc.SHUFFLE
+    )
     for column in columns:
         column.compress(root, compressor)
 
@@ -254,22 +274,28 @@ def check_format(root):
     if format_name != FORMAT_NAME:
         raise exceptions.FileFormatError(
             "Incorrect file format: expected '{}' got '{}'".format(
-                FORMAT_NAME, format_name))
+                FORMAT_NAME, format_name
+            )
+        )
     if format_version[0] < FORMAT_VERSION[0]:
         raise exceptions.FileFormatError(
             "Format version {} too old. Current version = {}".format(
-                format_version, FORMAT_VERSION))
+                format_version, FORMAT_VERSION
+            )
+        )
     if format_version[0] > FORMAT_VERSION[0]:
         raise exceptions.FileFormatError(
             "Format version {} too new. Current version = {}".format(
-                format_version, FORMAT_VERSION))
+                format_version, FORMAT_VERSION
+            )
+        )
 
 
 @contextlib.contextmanager
 def load_zarr(path):
     path = str(path)
     try:
-        store = zarr.ZipStore(path, mode='r')
+        store = zarr.ZipStore(path, mode="r")
     except zipfile.BadZipFile as bzf:
         raise exceptions.FileFormatError("File is not in tgzip format") from bzf
     root = zarr.group(store=store)
@@ -292,7 +318,8 @@ def decompress_zarr(root):
         location=root["individuals/location"],
         location_offset=root["individuals/location_offset"],
         metadata=root["individuals/metadata"],
-        metadata_offset=root["individuals/metadata_offset"])
+        metadata_offset=root["individuals/metadata_offset"],
+    )
 
     tables.nodes.set_columns(
         flags=root["nodes/flags"],
@@ -300,13 +327,15 @@ def decompress_zarr(root):
         population=root["nodes/population"],
         individual=root["nodes/individual"],
         metadata=root["nodes/metadata"],
-        metadata_offset=root["nodes/metadata_offset"])
+        metadata_offset=root["nodes/metadata_offset"],
+    )
 
     tables.edges.set_columns(
         left=coordinates[root["edges/left"]],
         right=coordinates[root["edges/right"]],
         parent=root["edges/parent"],
-        child=root["edges/child"])
+        child=root["edges/child"],
+    )
 
     tables.migrations.set_columns(
         left=coordinates[root["migrations/left"]],
@@ -314,14 +343,16 @@ def decompress_zarr(root):
         node=root["migrations/node"],
         source=root["migrations/source"],
         dest=root["migrations/dest"],
-        time=root["migrations/time"])
+        time=root["migrations/time"],
+    )
 
     tables.sites.set_columns(
         position=coordinates[root["sites/position"]],
         ancestral_state=root["sites/ancestral_state"],
         ancestral_state_offset=root["sites/ancestral_state_offset"],
         metadata=root["sites/metadata"],
-        metadata_offset=root["sites/metadata_offset"])
+        metadata_offset=root["sites/metadata_offset"],
+    )
 
     tables.mutations.set_columns(
         site=root["mutations/site"],
@@ -330,17 +361,20 @@ def decompress_zarr(root):
         derived_state=root["mutations/derived_state"],
         derived_state_offset=root["mutations/derived_state_offset"],
         metadata=root["mutations/metadata"],
-        metadata_offset=root["mutations/metadata_offset"])
+        metadata_offset=root["mutations/metadata_offset"],
+    )
 
     tables.populations.set_columns(
         metadata=root["populations/metadata"],
-        metadata_offset=root["populations/metadata_offset"])
+        metadata_offset=root["populations/metadata_offset"],
+    )
 
     tables.provenances.set_columns(
         timestamp=root["provenances/timestamp"],
         timestamp_offset=root["provenances/timestamp_offset"],
         record=root["provenances/record"],
-        record_offset=root["provenances/record_offset"])
+        record_offset=root["provenances/record_offset"],
+    )
 
     return tables.tree_sequence()
 
@@ -358,15 +392,15 @@ def print_summary(path, verbosity=0):
     arrays.sort(key=lambda x: x.nbytes_stored)
     max_name_len = max(len(array.name) for array in arrays)
     stored = [
-        humanize.naturalsize(array.nbytes_stored, binary=True) for array in arrays]
+        humanize.naturalsize(array.nbytes_stored, binary=True) for array in arrays
+    ]
     max_stored_len = max(len(size) for size in stored)
-    actual = [
-        humanize.naturalsize(array.nbytes, binary=True) for array in arrays]
+    actual = [humanize.naturalsize(array.nbytes, binary=True) for array in arrays]
     max_actual_len = max(len(size) for size in actual)
 
     line = "File: {}\t{}".format(
-        path,
-        humanize.naturalsize(os.path.getsize(path), binary=True))
+        path, humanize.naturalsize(os.path.getsize(path), binary=True)
+    )
     print(line)
     if verbosity > 0:
         print("format_version:", root.attrs["format_version"])
@@ -375,16 +409,28 @@ def print_summary(path, verbosity=0):
         print(json.dumps(prov, indent=4, sort_keys=True))
     fmt = "{:<{}} {:<{}}\t{:<{}}\t{}"
     line = fmt.format(
-        "name", max_name_len, "stored", max_stored_len, "actual", max_actual_len,
-        "ratio")
+        "name",
+        max_name_len,
+        "stored",
+        max_stored_len,
+        "actual",
+        max_actual_len,
+        "ratio",
+    )
     print(line)
     for array, stored, actual in zip(arrays, stored, actual):
         ratio = 0
         if array.nbytes > 0:
             ratio = array.nbytes_stored / array.nbytes
         line = fmt.format(
-            array.name, max_name_len, stored, max_stored_len, actual, max_actual_len,
-            "{:.2f}".format(ratio))
+            array.name,
+            max_name_len,
+            stored,
+            max_stored_len,
+            actual,
+            max_actual_len,
+            f"{ratio:.2f}",
+        )
         print(line)
         if verbosity > 0:
             for line in str(array.info).splitlines():
