@@ -1,21 +1,19 @@
-# -*- coding: utf-8 -*-
 #
 # Configuration file for the Sphinx documentation builder.
 #
 # This file does only contain a selection of the most common options. For a
 # full list see the documentation:
 # http://www.sphinx-doc.org/en/master/config
-
 # -- Path setup --------------------------------------------------------------
-
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
 import os
 import sys
-import pkg_resources
 from unittest.mock import MagicMock
+
+import pkg_resources
 
 
 class Mock(MagicMock):
@@ -34,17 +32,18 @@ MOCK_MODULES = [
 
 sys.modules.update((mod_name, Mock()) for mod_name in MOCK_MODULES)
 
-sys.path.insert(0, os.path.abspath('../'))
+sys.path.insert(0, os.path.abspath("../"))
 
 # -- Project information -----------------------------------------------------
 
-project = 'tszip'
-copyright = '2019, Tskit Developers'
-author = 'Tskit Developers'
+project = "tszip"
+copyright = "2019, Tskit Developers"  # noqa: A001
+author = "Tskit Developers"
 
 try:
     from setuptools_scm import get_version
-    release = get_version(root='..', relative_to=__file__)
+
+    release = get_version(root="..", relative_to=__file__)
     version = release[:3]
 except pkg_resources.DistributionNotFound:
     release = "0.0.0"
@@ -61,23 +60,23 @@ except pkg_resources.DistributionNotFound:
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.intersphinx',
-    'sphinx.ext.viewcode',
-    'sphinxarg.ext',
+    "sphinx.ext.autodoc",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.viewcode",
+    "sphinxarg.ext",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = '.rst'
+source_suffix = ".rst"
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -89,7 +88,7 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The name of the Pygments (syntax highlighting) style to use.
 pygments_style = None
@@ -100,7 +99,7 @@ pygments_style = None
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = 'sphinx_rtd_theme'
+html_theme = "sphinx_rtd_theme"
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
@@ -111,7 +110,7 @@ html_theme = 'sphinx_rtd_theme'
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
 
 # Custom sidebar templates, must be a dictionary that maps document names
 # to template names.
@@ -127,7 +126,7 @@ html_static_path = ['_static']
 # -- Options for HTMLHelp output ---------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = 'tszipdoc'
+htmlhelp_basename = "tszipdoc"
 
 
 # -- Options for LaTeX output ------------------------------------------------
@@ -136,15 +135,12 @@ latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
     # 'papersize': 'letterpaper',
-
     # The font size ('10pt', '11pt' or '12pt').
     #
     # 'pointsize': '10pt',
-
     # Additional stuff for the LaTeX preamble.
     #
     # 'preamble': '',
-
     # Latex figure (float) alignment
     #
     # 'figure_align': 'htbp',
@@ -154,8 +150,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, 'tszip.tex', 'tszip Documentation',
-     'Tskit Developers', 'manual'),
+    (master_doc, "tszip.tex", "tszip Documentation", "Tskit Developers", "manual"),
 ]
 
 
@@ -163,10 +158,7 @@ latex_documents = [
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'tszip', 'tszip Documentation',
-     [author], 1)
-]
+man_pages = [(master_doc, "tszip", "tszip Documentation", [author], 1)]
 
 
 # -- Options for Texinfo output ----------------------------------------------
@@ -175,9 +167,15 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, 'tszip', 'tszip Documentation',
-     author, 'tszip', 'One line description of project.',
-     'Miscellaneous'),
+    (
+        master_doc,
+        "tszip",
+        "tszip Documentation",
+        author,
+        "tszip",
+        "One line description of project.",
+        "Miscellaneous",
+    ),
 ]
 
 
@@ -196,7 +194,7 @@ epub_title = project
 # epub_uid = ''
 
 # A list of files that should not be packed into the epub file.
-epub_exclude_files = ['search.html']
+epub_exclude_files = ["search.html"]
 
 
 # -- Extension configuration -------------------------------------------------
@@ -205,6 +203,6 @@ epub_exclude_files = ['search.html']
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
-    'https://docs.python.org/3/': None,
-    'tskit': ('https://tskit.readthedocs.io/en/stable', None),
+    "https://docs.python.org/3/": None,
+    "tskit": ("https://tskit.readthedocs.io/en/stable", None),
 }
